pyproject.toml
[project]
name = "axiom-spec"
version = "0.1.0"
description = "Compile .axiom specs into verified Python code using LLMs. Specs are source, code is cache."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ryan Colkitt", email = "ryan@vaspera.com"}
]
keywords = [
    "llm",
    "code-generation",
    "specification",
    "verification",
    "ai",
    "claude",
    "testing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "litellm>=1.0.0",
    "anthropic>=0.18.0",
    "structlog>=24.0.0",
    "tomli>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "pygls>=1.3.0",
    "lsprotocol>=2024.0.0",
    "watchdog>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/RCOLKITT/Axiom"
Documentation = "https://github.com/RCOLKITT/Axiom#readme"
Repository = "https://github.com/RCOLKITT/Axiom"
Issues = "https://github.com/RCOLKITT/Axiom/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
axiom = "axiom.cli.main:cli"
axiom-lsp = "axiom.lsp.server:start_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axiom"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
    "SIM102", # nested if statements (often clearer than combined)
    "SIM108", # ternary operator (sometimes less readable)
]

[tool.ruff.lint.isort]
known-first-party = ["axiom"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src/axiom"]

[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "hypothesis.*",
    "starlette.*",
    "pygls.*",
    "lsprotocol.*",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
